Refactor classes into modules
I had a POC working, but the design was not great.
Adding new ways to extract data would mean modifying two classes,
and both would grow and mix concerns, esp if some methods were
more complicated.

This way just requires a mixin to be added to the main Parser class.
This could be improved more if the list grew by using
IoC/dry-containers.